tweaked crime stats
- moved functions into crimeStats.js
- formatted the crime category names
- somehow the css changes disappeared on my end so i added those back
